# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  requests.CreateAddressRequest:
    properties:
      city:
        example: San Francisco
        type: string
      country:
        example: USA
        type: string
      latitude:
        example: 37.7749
        type: number
      longitude:
        example: -122.4194
        type: number
      postal_code:
        example: "94105"
        type: string
      state:
        example: CA
        type: string
      street:
        example: 123 Main St
        type: string
    type: object
...
  responses.AddressResponse:
    description: Address response model
    properties:
      city:
        example: San Francisco
        type: string
      country:
        example: USA
        type: string
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      latitude:
        example: 37.7749
        type: number
      longitude:
        example: -122.4194
        type: number
      postal_code:
        example: "94105"
        type: string
      state:
        example: CA
        type: string
      street:
        example: 123 Main St
        type: string
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
    type: object
  responses.BaseResponse:
    description: Standard response structure for all API endpoints
    properties:
      currentPage:
        type: integer
      data: {}
      details: {}
      error: {}
      lastPage:
        type: integer
      message: {}
      perPage:
        type: integer
      total:
        type: integer
    type: object
...
info:
  contact: {}
  description: This is a API Server.
  title: DockMaster Web API
  version: 0.0.1
paths:
  /addresses:
    post:
      consumes:
      - application/json
      description: Creates a new address in the system
      parameters:
      - description: Address details
        in: body
        name: address
        required: true
        schema:
          $ref: '#/definitions/requests.CreateAddressRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responses.AddressResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: Create address
      tags:
      - Addresses
  /addresses/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves an address by its ID
      parameters:
      - description: Address ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.AddressResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: Get address by ID
      tags:
      - Addresses
    put:
      consumes:
      - application/json
      description: Updates an existing address
      parameters:
      - description: Address ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Address details
        in: body
        name: address
        required: true
        schema:
          $ref: '#/definitions/requests.UpdateAddressRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.AddressResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: Update address
      tags:
      - Addresses
  /auth/login:
    post:
      consumes:
      - application/json
      description: Perform user login
      operationId: user-login
      parameters:
      - description: User's credentials
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success response with login data
          schema:
            $ref: '#/definitions/responses.LoginResponseWrapper'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/responses.Error'
        "401":
          description: Authentication error
          schema:
            $ref: '#/definitions/responses.Error'
        "403":
          description: Account locked
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Authenticate a user
      tags:
      - Authentication
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Perform refresh access token
      operationId: user-refresh
      parameters:
      - description: Refresh token
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success response with new tokens
          schema:
            $ref: '#/definitions/responses.LoginResponseWrapper'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/responses.Error'
        "401":
          description: Authentication error
          schema:
            $ref: '#/definitions/responses.Error'
        "403":
          description: Account locked
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Refresh access token
      tags:
      - Authentication
  /boats/create:
    post:
      consumes:
      - application/json
      description: Creates a new boat
      parameters:
      - description: Boat information
        in: body
        name: boat
        required: true
        schema:
          $ref: '#/definitions/requests.BoatCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.BoatResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Create boat
      tags:
      - Boats
  /boats/customer:
    get:
      consumes:
      - application/json
      description: Retrieves boats associated with a customer
      parameters:
      - description: Customer ID
        in: query
        name: CustomerId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.BoatSearchResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Retrieve boats for customer
      tags:
      - Boats
  /boats/list:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated list of boats
      parameters:
      - description: Page number
        in: query
        minimum: 1
        name: page
        required: true
        type: integer
      - description: Page size
        in: query
        maximum: 100
        minimum: 1
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.BoatListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: List boats by page
      tags:
      - Boats
  /boats/retrieve:
    get:
      consumes:
      - application/json
      description: Retrieves a boat by its ID
      parameters:
      - description: Boat ID
        in: query
        name: BoatId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.BoatResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Retrieve boat by ID
      tags:
      - Boats
  /boats/search:
    get:
      consumes:
      - application/json
      description: Searches for boats based on search string
      parameters:
      - description: Search string
        in: query
        name: SearchString
        required: true
        type: string
      - default: false
        description: Direct hit search
        in: query
        name: DirectHit
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.BoatSearchResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Search boats
      tags:
      - Boats
  /boats/update:
    post:
      consumes:
      - application/json
      description: Updates a boat's information
      parameters:
      - description: Boat information
        in: body
        name: boat
        required: true
        schema:
          $ref: '#/definitions/dme.BoatUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.BoatResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Update boat
      tags:
      - Boats
  /customers/create:
    post:
      consumes:
      - application/json
      description: Creates a new customer
      parameters:
      - description: Customer information
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/requests.CustomerCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.CustomerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Create customer
      tags:
      - Customers
  /customers/list:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated list of customers
      parameters:
      - description: Page number
        in: query
        minimum: 1
        name: page
        required: true
        type: integer
      - description: Page size
        in: query
        maximum: 100
        minimum: 1
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.CustomerListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: List customers by page
      tags:
      - Customers
  /customers/list-short:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated short list of customers
      parameters:
      - description: Page number
        in: query
        minimum: 1
        name: page
        required: true
        type: integer
      - description: Page size
        in: query
        maximum: 100
        minimum: 1
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.CustomerListShortResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: List customers short
      tags:
      - Customers
  /customers/retrieve:
    get:
      consumes:
      - application/json
      description: Retrieves a customer by their ID
      parameters:
      - description: Customer ID
        in: query
        name: CustomerId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.CustomerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Retrieve customer by ID
      tags:
      - Customers
  /customers/search:
    get:
      consumes:
      - application/json
      description: Searches for customers based on search string
      parameters:
      - description: Search string
        in: query
        name: SearchString
        required: true
        type: string
      - default: false
        description: Direct hit search
        in: query
        name: DirectHit
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.CustomerSearchResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Search customers
      tags:
      - Customers
  /customers/settings:
    get:
      consumes:
      - application/json
      description: Retrieves settings for a customer, creates with defaults if not
        found
      parameters:
      - description: Customer ID
        in: query
        name: CustomerId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.CustomerSettingsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Get customer settings
      tags:
      - Customers
    post:
      consumes:
      - application/json
      description: Updates settings for a customer
      parameters:
      - description: Customer settings
        in: body
        name: settings
        required: true
        schema:
          $ref: '#/definitions/requests.CustomerSettingsUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.CustomerSettingsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Update customer settings
      tags:
      - Customers
  /customers/update:
    post:
      consumes:
      - application/json
      description: Updates a customer's information
      parameters:
      - description: Customer information
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/dme.CustomerUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.CustomerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Update customer
      tags:
      - Customers
  /dme/credentials:
    post:
      consumes:
      - application/json
      description: Create new DME API credentials for an organization
      parameters:
      - description: Create DME credential request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.CreateDMECredentialRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.DMECredentialResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Create DME credentials
      tags:
      - DME Credentials
...
  /dme/sysids:
    get:
      consumes:
      - application/json
      description: Get a list of all DME System IDs
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.DMESysIDListResponseWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: List all DME System IDs
      tags:
      - DME System IDs
    post:
      consumes:
      - application/json
      description: Create a new DME System ID for an organization
      parameters:
      - description: Create DME System ID request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.CreateDMESysIDRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.DMESysIDResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Create DME System ID
      tags:
      - DME System IDs
...
  /documents/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a document
      parameters:
      - description: Document ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete document
      tags:
      - Documents
    get:
      consumes:
      - application/json
      description: Retrieves a specific document by ID
      parameters:
      - description: Document ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/responses.DocumentResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: Get document
      tags:
      - Documents
...
  /email/send-html:
    post:
      consumes:
      - application/json
      description: Send an email with HTML content
      parameters:
      - description: Email details
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.SendHTMLEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Email accepted for delivery
          schema:
            $ref: '#/definitions/responses.EmailSendResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Send HTML email
      tags:
      - Email
...
  /user/profile:
    get:
      consumes:
      - application/json
      description: get my user
      produces:
      - application/json
      responses:
        "200":
          description: Current user's profile
          schema:
            $ref: '#/definitions/responses.UserResponseWrapper'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/responses.Error'
      security:
      - ApiKeyAuth: []
      summary: Get my user
      tags:
      - User
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"